// apps/mobile/lib/types.ts
export type ChatMessage = {
	id: number | string;
	roomId: string;
	chatDocId?: string;
	type: 'text' | 'image' | 'gif' | 'file' | 'audio' | 'video';
	chatInfo: string;
	fileName?: string;
	isMsgEdited?: boolean;
	isMsgSaved?: boolean;
	isAIMessage?: boolean;
	reactions?: Array<{
		id: string;
		reactors: Array<{ uid: string; name: string }>;
	}>;
	userUid: string;
	userName: string;
	userPhoto: string;
	time: any; //TODO: fix
	isUserInfoDisplayed?: boolean;
	isConsecutiveMessage?: boolean;
	isDate?: boolean;
};